third_party/WebKit/LayoutTests/virtual/gpu/fast/canvas/canvas-toDataURL-case-insensitive-mimetype-expected.txt

Test that toDataURL(mimeType) ignores the case of 'mimeType'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tryMimeType('image/png') is 'image/png'
PASS tryMimeType('iMAge/Png') is 'image/png'
PASS tryMimeType('IMAGE/PNG') is 'image/png'
PASS tryMimeType('image/jpeg') is 'image/jpeg'
PASS tryMimeType('imAgE/jPEg') is 'image/jpeg'
PASS tryMimeType('IMAGE/JPEG') is 'image/jpeg'
PASS tryMimeType('image/webp') is 'image/webp'
PASS tryMimeType('imAgE/wEBp') is 'image/webp'
PASS tryMimeType('IMAGE/WEBP') is 'image/webp'
PASS successfullyParsed is true

TEST COMPLETE
